+++ src/components/CreatePlanRemindMessage/CreatePlanRemindMessage.tsx
@@ -1,11 +1,9 @@
 'use client';
 
 import { ModalSendRemindExample, WritableRemindItem } from '@/components';
-import { SESSION_STORAGE_KEY } from '@/constants';
-import { useSessionStorage } from '@/hooks/useSessionStorage';
-import { RemindItemType } from '@/types';
 import classNames from 'classnames';
-import React, { useCallback, useEffect, useState } from 'react';
+import React from 'react';
+import useCreatePlanRemindMessage from './hooks/useCreatePlanRemindMessage';
 import './index.scss';
 
 interface CreatePlanRemindMessageProps {
@@ -17,53 +15,16 @@ export default function CreatePlanRemindMessage({
   setIsLastStepDataAllExist,
   isCreateOrEditPage,
 }: CreatePlanRemindMessageProps) {
-  const [remindMessageList, setRemindMessageList] = useSessionStorage<
-    RemindItemType[]
-  >({
-    key:
-      isCreateOrEditPage === 'create'
-        ? SESSION_STORAGE_KEY.STEP_4
-        : SESSION_STORAGE_KEY.EDIT_REMIND_MESSAGE,
-    initialValue: [],
+  const {
+    handleChangeRemindMessage,
+    makeAllRemindMessageSame,
+    remindMessageList,
+    isSendRemindModalOpen,
+    setIsSendRemindModalOpen,
+  } = useCreatePlanRemindMessage({
+    setIsLastStepDataAllExist,
+    isCreateOrEditPage,
   });
-
-  useEffect(() => {
-    if (remindMessageList) {
-      if (
-        remindMessageList.every((remindItem) => remindItem.message.length > 0)
-      ) {
-        setIsLastStepDataAllExist(true);
-      } else {
-        setIsLastStepDataAllExist(false);
-      }
-    }
-  }, [remindMessageList, setIsLastStepDataAllExist]);
-
-  const handleChangeRemindMessage = (
-    month: number,
-    day: number,
-    newMessage: string,
-  ) => {
-    setRemindMessageList(
-      remindMessageList.map((item) => {
-        if (item.date.month === month && item.date.day === day) {
-          return { ...item, message: newMessage };
-        }
-        return item;
-      }),
-    );
-  };
-
-  const makeAllRemindMessageSame = useCallback(() => {
-    if (remindMessageList.length > 1) {
-      const firstMessage = remindMessageList[0].message;
-      setRemindMessageList(
-        remindMessageList.map((item) => ({ ...item, message: firstMessage })),
-      );
-    }
-  }, [remindMessageList, setRemindMessageList]);
-
-  const [isSendRemindModalOpen, setIsSendRemindModalOpen] = useState(false);
 
   return (
     <div className={classNames(['create-remind-message'])}>
